implement isa and refactor tsr

+++ android-app/app/src/main/java/com/copilotovirtual/data/model/SpeedLimitState.kt
@@ -2,5 +2,5 @@ package com.copilotovirtual.data.model
 
 object SpeedLimitState {
     // Máxima velocidad permitida en la ciudad
-    var currentSpeedLimit: Float = 40.0f
+    var currentSpeedLimit: Int = 40
 }
