+++ android-app/app/src/main/java/com/copilotovirtual/data/model/SpeedState.kt
@@ -1,5 +1,5 @@
 package com.copilotovirtual.data.model
 
 object SpeedState {
-    var currentSpeed: Float = 0.0f
+    var currentSpeed: Int = 0
 }
